# backend/docs/swagger.yml
openapi: '3.0.3'
info:
  title: 'GRIND API'
  version: '0.0.6'
...
paths:
...
  /users/me/invite-codes:
    get:
      tags: [User]
      summary: 'Invite codes of authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      responses:
        '200':
          $ref: '#/components/responses/GetMyInviteCodes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []

    post:
      tags: [User]
      summary: 'Create invite code by authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      requestBody:
        $ref: '#/components/requestBodies/CreateInviteCode'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      security:
        - Bearer: []
  /users/me/profile:
    get:
      tags: [User]
      summary: 'Get profile of authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      responses:
        '200':
          $ref: '#/components/responses/GetMyProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []
  /users/me/skills:
    get:
      tags: [User]
      summary: 'Get skills of authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      responses:
        '200':
          $ref: '#/components/responses/GetMySkills'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []
  /users/me/skills-progress:
    get:
      tags: [User]
      summary: 'Get skills progress of authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      responses:
        '200':
          $ref: '#/components/responses/GetMySkillsProgress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []
  /users/me/activities:
    get:
      tags: [User]
      summary: 'Get activities of authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
        - $ref: '#/components/parameters/PaginationPageQuery'
        - $ref: '#/components/parameters/PaginationLimitQuery'

      responses:
        '200':
          $ref: '#/components/responses/GetMyActivities'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - Bearer: []

    post:
      tags: [User]
      summary: 'Create activity for authorized user'
      parameters:
        - $ref: '#/components/parameters/JWTAuthToken'
      requestBody:
        $ref: '#/components/requestBodies/CreateActivity'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
      security:
        - Bearer: []
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    JWTAuthToken:
      name: Authorization
      in: header
      required: true
      schema: { type: string }
      example: 'Bearer xxxxxxxxxxx'
    PaginationPageQuery:
      name: page
      in: query
      required: true
      schema: { type: integer, default: 1, minimum: 1 }
    PaginationLimitQuery:
      name: limit
      in: query
      required: true
      schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
  schemas:
    SuccessResponse:
      type: object
      properties:
        data:
          nullable: true
    PaginationResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - properties:
            data:
              type: object
              properties:
                pagination:
                  type: object
                  properties:
                    total_pages: { type: integer }
                    current_page: { type: integer }
                    next_page: { type: integer }
                    limit: { type: integer }
                items:
                  type: array
                  items:
                    nullable: true
    ErrorsResponse:
      type: object
      properties:
        errors:
          type: object
          additionalProperties: { type: string }
    UUID:
      type: string
      format: uuid
    DateTime:
      type: string
      format: datetime
    URL:
      type: string
      format: uri
    BaseInviteCode:
      type: object
      properties:
        code: { type: string }
        max_uses: { type: integer }
    InviteCodeListItem:
      allOf:
        - $ref: '#/components/schemas/BaseInviteCode'
        - properties:
            id: { $ref: '#/components/schemas/UUID' }
            uses: { type: integer }
            created_at: { $ref: '#/components/schemas/DateTime' }
    User:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        telegram_id: { type: integer }
        name: { type: string }
        avatar_url: { $ref: '#/components/schemas/URL' }
        level: { type: integer }
        created_at: { $ref: '#/components/schemas/DateTime' }
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - properties:
            inviter: { $ref: '#/components/schemas/User' }
    SkillProgress:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        total_xp: { type: integer }
        current_level: { type: integer }
        next_level: { type: integer }
        current_level_start_xp: { type: integer }
        next_level_start_xp: { type: integer }
        xp_to_next_level: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/SkillProgress'
    Skill:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
    ActivityReward:
      type: object
      properties:
        skill_id: { $ref: '#/components/schemas/UUID' }
        skill_name: { type: string }
        xp_amount: { type: integer }
    ActivityType:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
    ActivityListItem:
      type: object
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        description: { type: string }
        activity_type: { $ref: '#/components/schemas/ActivityType' }
        has_impact: { type: boolean }
        is_hard: { type: boolean }
        is_new: { type: boolean }
        rewards:
          type: array
          items:
            $ref: '#/components/schemas/ActivityReward'
        created_at: { $ref: '#/components/schemas/DateTime' }
    BaseCreatedActivity:
      type: object
      properties:
        mode: { type: string, enum: ['manually', 'ai'] }
        description: { type: string }
    SkillWeight:
      type: object
      properties:
        skill_id: { $ref: '#/components/schemas/UUID' }
        weight: { type: number }
    ManuallyCreatedActivity:
      allOf:
        - $ref: '#/components/schemas/BaseCreatedActivity'
        - properties:
            activity_type_id: { $ref: '#/components/schemas/UUID' }
            has_impact: { type: boolean }
            is_new: { type: boolean }
            is_hard: { type: boolean }
            skill_weights:
              type: array
              items:
                $ref: '#/components/schemas/SkillWeight'
    AICreatedActivity:
      allOf:
        - $ref: '#/components/schemas/BaseCreatedActivity'
...
    CreateInviteCodeRequestBody:
      $ref: '#/components/schemas/BaseInviteCode'
    GetMyInviteCodesResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/InviteCodeListItem'
    GetMyProfileResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - properties:
            data:
              $ref: '#/components/schemas/UserProfile'
    GetMySkillsProgressResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/SkillProgress'
    CreateActivityRequestBody:
      oneOf:
        - $ref: '#/components/schemas/ManuallyCreatedActivity'
        - $ref: '#/components/schemas/AICreatedActivity'
    GetMyActivitiesResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationResponse'
        - properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ActivityListItem'
...
    GetMySkillsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
...
  examples:
    EmptySuccessResponse:
      value:
        data: []
    BadRequestResponse:
      value:
        errors:
          other: 'Something went wrong'
    UnauthorizedResponse:
      value:
        errors:
          other: 'Unauthorized'
...
    UnprocessableEntityResponse:
      value:
        errors:
          field1: 'field1 is required'
          field2: 'field2 has invalid format'
          field3: 'field3 must be numeric'
...
    CreateInviteCodeRequestBody:
      value:
        code: 'INVITE12345'
        max_uses: 5
    GetMyInviteCodesResponse:
      value:
        data:
          - id: '550e8400-e29b-41d4-a716-446655440000'
            code: 'INVITE12345'
            uses: 2
            max_uses: 5
            created_at: '2024-01-01T12:00:00Z'
          - id: '550e8400-e29b-41d4-a716-446655440001'
            code: 'INVITE67890'
            uses: 0
            max_uses: 3
            created_at: '2024-01-02T15:30:00Z'
    GetMyProfileResponse:
      value:
        data:
          id: '550e8400-e29b-41d4-a716-446655440002'
          telegram_id: 10000323233
          name: 'UF2xx'
          avatar_url: 'https://example.com/avatar.jpg'
          level: 3
          created_at: '2024-01-02T15:30:00Z'
          inviter:
            id: '550e8400-e29b-41d4-a716-446655440003'
            telegram_id: 10000543532
            name: 'sunsetsavorer'
            avatar_url: 'https://example.com/inviter_avatar.jpg'
            level: 5
            created_at: '2024-01-02T15:30:00Z'
    GetMySkillsProgressResponse:
      value:
        data:
          - id: '550e8400-e29b-41d4-a716-446655440004'
            name: 'Endurance'
            total_xp: 1500
            current_level: 3
            next_level: 4
            current_level_start_xp: 1000
            next_level_start_xp: 2000
            xp_to_next_level: 500
            items:
              - id: '550e8400-e29b-41d4-a716-446655440005'
                name: 'Running'
                total_xp: 800
                current_level: 2
                next_level: 4
                current_level_start_xp: 500
                next_level_start_xp: 1500
                xp_to_next_level: 700
              - id: '550e8400-e29b-41d4-a716-446655440006'
                name: 'Swimming'
                total_xp: 700
                current_level: 2
                next_level: 3
                current_level_start_xp: 300
                next_level_start_xp: 1200
                xp_to_next_level: 500
          - id: '550e8400-e29b-41d4-a716-446655440005'
            name: 'Power'
            total_xp: 500
            current_level: 2
            next_level: 3
            current_level_start_xp: 300
            next_level_start_xp: 800
            xp_to_next_level: 300
            items:
              - id: '550e8400-e29b-41d4-a716-446655440006'
                name: 'Climbing'
                total_xp: 500
                current_level: 2
                next_level: 3
                current_level_start_xp: 300
                next_level_start_xp: 800
                xp_to_next_level: 300
          - id: '550e8400-e29b-41d4-a716-446655440006'
            name: 'Intellect'
            total_xp: 200
            current_level: 1
            next_level: 2
            current_level_start_xp: 0
            next_level_start_xp: 500
            xp_to_next_level: 300
            items:
              - id: '550e8400-e29b-41d4-a716-446655440007'
                name: 'Programming'
                total_xp: 200
                current_level: 1
                next_level: 2
                current_level_start_xp: 0
                next_level_start_xp: 500
                xp_to_next_level: 300
    ManuallyCreatedActivityRequestBody:
      value:
        mode: 'manually'
        description: 'Do 10 pushups'
        activity_type_id: '550e8400-e29b-41d4-a716-446655440008'
        has_impact: true
        is_new: true
        is_hard: false
        skill_weights:
          - skill_id: '550e8400-e29b-41d4-a716-446655440004'
            weight: 0.3
          - skill_id: '550e8400-e29b-41d4-a716-446655440005'
            weight: 0.7
    AICreatedActivityRequestBody:
      value:
        mode: 'ai'
        description: 'Run 5 kilometers'
    GetMyActivitiesResponse:
      value:
        data:
          pagination:
            total_pages: 5
            current_page: 1
            next_page: 2
            limit: 10
          items:
            - id: '550e8400-e29b-41d4-a716-446655440000'
              description: 'Do 10 pushups'
              activity_type:
                id: '550e8400-e29b-41d4-a716-446655440008'
                name: 'Physical'
              has_impact: true
              is_hard: false
              is_new: true
              rewards:
                - skill_id: '550e8400-e29b-41d4-a716-446655440004'
                  skill_name: 'Endurance'
                  xp_amount: 100
                - skill_id: '550e8400-e29b-41d4-a716-446655440005'
                  skill_name: 'Power'
                  xp_amount: 200
              created_at: '2024-01-02T15:30:00Z'
            - id: '550e8400-e29b-41d4-a716-446655440001'
              description: 'Deploy app in kubernetes'
              activity_type:
                id: '550e8400-e29b-41d4-a716-446655440009'
                name: 'Mental'
              has_impact: true
              is_hard: true
              is_new: false
              rewards:
                - skill_id: '550e8400-e29b-41d4-a716-446655440006'
                  skill_name: 'Intellect'
                  xp_amount: 300
              created_at: '2024-01-03T10:00:00Z'
...
    GetMySkillsResponse:
      value:
        data:
          - id: '550e8400-e29b-41d4-a716-446655440004'
            name: 'Endurance'
          - id: '550e8400-e29b-41d4-a716-446655440005'
            name: 'Power'
          - id: '550e8400-e29b-41d4-a716-446655440006'
            name: 'Intellect'
...
  requestBodies:
...
    CreateInviteCode:
      description: 'Request body for creating invite code'
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateInviteCodeRequestBody'
          examples:
            base:
              $ref: '#/components/examples/CreateInviteCodeRequestBody'
    CreateActivity:
      description: 'Request body for creating activity'
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateActivityRequestBody'
          examples:
            manually:
              $ref: '#/components/examples/ManuallyCreatedActivityRequestBody'
            ai:
              $ref: '#/components/examples/AICreatedActivityRequestBody'
...
  responses:
    EmptySuccess:
      description: 'Success response'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
          examples:
            success:
              $ref: '#/components/examples/EmptySuccessResponse'
    BadRequest:
      description: 'Something went wrong'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsResponse'
          examples:
            error:
              $ref: '#/components/examples/BadRequestResponse'
    Unauthorized:
      description: 'Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsResponse'
          examples:
            error:
              $ref: '#/components/examples/UnauthorizedResponse'
...
    UnprocessableEntity:
      description: 'Validation error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsResponse'
          examples:
            error:
              $ref: '#/components/examples/UnprocessableEntityResponse'
...
    GetMyInviteCodes:
      description: 'Response for getting invite codes of authorized user'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetMyInviteCodesResponse'
          examples:
            success:
              $ref: '#/components/examples/GetMyInviteCodesResponse'
    GetMyProfile:
      description: 'Response for getting profile of authorized user'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetMyProfileResponse'
          examples:
            success:
              $ref: '#/components/examples/GetMyProfileResponse'
    GetMySkillsProgress:
      description: 'Response for getting skills of authorized user'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetMySkillsProgressResponse'
          examples:
            success:
              $ref: '#/components/examples/GetMySkillsProgressResponse'
    GetMyActivities:
      description: 'Response for getting activities of authorized user'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetMyActivitiesResponse'
          examples:
            success:
              $ref: '#/components/examples/GetMyActivitiesResponse'
...
    GetMySkills:
      description: 'Response for getting skills of authorized user'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetMySkillsResponse'
          examples:
            success:
              $ref: '#/components/examples/GetMySkillsResponse'