# backend/internal/models/activity_type.go
package models

import (
	"time"

	"github.com/google/uuid"
)

type ActivityType struct {
	ID        uuid.UUID `gorm:"column:id;primaryKey;default:gen_random_uuid()" json:"id"`
	Name      string    `gorm:"column:name" json:"name"`
	CreatedAt time.Time `gorm:"column:created_at" json:"created_at"`
}